requirements-test.txt
pytest
requests_mock